refactor: TS in Ingredient
